+++ Sources/Base/OAuth2Error.swift
@@ -30,6 +30,7 @@ public enum OAuth2Error: ErrorType, CustomStringConvertible, Equatable {
 	case Generic(String)
 	case NSError(Foundation.NSError)
 	
+	// Client errors
 	case NoClientId
 	case NoClientSecret
 	case NoRedirectURL
@@ -42,6 +43,11 @@ public enum OAuth2Error: ErrorType, CustomStringConvertible, Equatable {
 	
 	case NoRegistrationURL
 	
+	// Request errors
+	case NotUsingTLS
+	case UnableToOpenAuthorizeURL
+	case InvalidRequest
+	case RequestCancelled
 	case NoTokenType
 	case UnsupportedTokenType(String)
 	case NoDataInResponse
@@ -51,9 +57,7 @@ public enum OAuth2Error: ErrorType, CustomStringConvertible, Equatable {
 	case UTF8EncodeError
 	case UTF8DecodeError
 	
-	case NotUsingTLS
-	case UnableToOpenAuthorizeURL
-	case InvalidRequest
+	// OAuth2 errors
 	case UnauthorizedClient
 	case AccessDenied
 	case UnsupportedResponseType
@@ -112,12 +116,20 @@ public enum OAuth2Error: ErrorType, CustomStringConvertible, Equatable {
 		case .NoRegistrationURL:
 			return "No registration URL defined"
 		
+		case .NotUsingTLS:
+			return "You MUST use HTTPS/SSL/TLS"
+		case .UnableToOpenAuthorizeURL:
+			return "Cannot open authorize URL"
+		case .InvalidRequest:
+			return "The request is missing a required parameter, includes an invalid parameter value, includes a parameter more than once, or is otherwise malformed."
+		case .RequestCancelled:
+			return "The request has been cancelled"
 		case NoTokenType:
 			return "No token type received, will not use the token"
 		case UnsupportedTokenType(let message):
 			return message
 		case NoDataInResponse:
-			return "No date in the response"
+			return "No data in the response"
 		case PrerequisiteFailed(let message):
 			return message
 		case InvalidState:
@@ -129,12 +141,6 @@ public enum OAuth2Error: ErrorType, CustomStringConvertible, Equatable {
 		case UTF8DecodeError:
 			return "Failed to decode given data as a UTF-8 string"
 		
-		case .NotUsingTLS:
-			return "You MUST use HTTPS/SSL/TLS"
-		case .UnableToOpenAuthorizeURL:
-			return "Cannot open authorize URL"
-		case .InvalidRequest:
-			return "The request is missing a required parameter, includes an invalid parameter value, includes a parameter more than once, or is otherwise malformed."
 		case .UnauthorizedClient:
 			return "The client is not authorized to request an access token using this method."
 		case .AccessDenied:
@@ -169,6 +175,10 @@ public func ==(lhs: OAuth2Error, rhs: OAuth2Error) -> Bool {
 	case (.InvalidRedirectURL(let lhu), .InvalidRedirectURL(let rhu)):       return lhu == rhu
 	case (.NoRefreshToken, .NoRefreshToken):			         return true
 	
+	case (.NotUsingTLS, .NotUsingTLS):                           return true
+	case (.UnableToOpenAuthorizeURL, .UnableToOpenAuthorizeURL): return true
+	case (.InvalidRequest, .InvalidRequest):                     return true
+	case (.RequestCancelled, .RequestCancelled):                 return true
 	case (.NoTokenType, .NoTokenType):                           return true
 	case (.UnsupportedTokenType(let lhm), .UnsupportedTokenType(let rhm)):   return lhm == rhm
 	case (.NoDataInResponse, .NoDataInResponse):                 return true
@@ -178,9 +188,6 @@ public func ==(lhs: OAuth2Error, rhs: OAuth2Error) -> Bool {
 	case (.UTF8EncodeError, .UTF8EncodeError):                   return true
 	case (.UTF8DecodeError, .UTF8DecodeError):                   return true
 	
-	case (.NotUsingTLS, .NotUsingTLS):                           return true
-	case (.UnableToOpenAuthorizeURL, .UnableToOpenAuthorizeURL): return true
-	case (.InvalidRequest, .InvalidRequest):                     return true
 	case (.UnauthorizedClient, .UnauthorizedClient):             return true
 	case (.AccessDenied, .AccessDenied):                         return true
 	case (.UnsupportedResponseType, .UnsupportedResponseType):   return true